tests: Split non-property tests out of test_*_properties.py
