+++ tests/component/system/storage/test_persisted_scale.py
@@ -1,0 +1,16 @@
+from nidaqmx.system.storage import PersistedScale
+
+
+def test___persisted_scales_with_same_name___compare___equal(init_kwargs):
+    persisted_scale1 = PersistedScale("Scale1", **init_kwargs)
+    persisted_scale2 = PersistedScale("Scale1", **init_kwargs)
+
+    assert persisted_scale1 is not persisted_scale2
+    assert persisted_scale1 == persisted_scale2
+
+
+def test___persisted_scales_with_different_names___compare___not_equal(init_kwargs):
+    persisted_scale1 = PersistedScale("Scale1", **init_kwargs)
+    persisted_scale2 = PersistedScale("Scale2", **init_kwargs)
+
+    assert persisted_scale1 != persisted_scale2
