+++ tests/component/system/storage/test_persisted_scale_properties.py
@@ -22,21 +22,6 @@ def test___nonexistent_persisted_scale___get_property___raises_custom_scale_does
     assert exc_info.value.error_code == DAQmxErrors.CUSTOM_SCALE_DOES_NOT_EXIST
 
 
-def test___persisted_scales_with_same_name___compare___equal(init_kwargs):
-    persisted_scale1 = PersistedScale("Scale1", **init_kwargs)
-    persisted_scale2 = PersistedScale("Scale1", **init_kwargs)
-
-    assert persisted_scale1 is not persisted_scale2
-    assert persisted_scale1 == persisted_scale2
-
-
-def test___persisted_scales_with_different_names___compare___not_equal(init_kwargs):
-    persisted_scale1 = PersistedScale("Scale1", **init_kwargs)
-    persisted_scale2 = PersistedScale("Scale2", **init_kwargs)
-
-    assert persisted_scale1 != persisted_scale2
-
-
 @pytest.mark.scale_name("double_gain_scale")
 def test___persisted_scale___get_bool_property___returns_persisted_value(persisted_scale):
     assert persisted_scale.allow_interactive_editing
